backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
paddleocr==2.7.3
paddlepaddle==2.6.2
Pillow==10.1.0
numpy==1.26.4
